bug: transfer-pvc clustertask's bad default value (#41)
The default value for transfer-pvc's `optional-flags` argument is not
correct. After playing with it offline, it does not appear there is a
string that can be passed to `optional-flags` that will be accepted.

Examples include `--optional-flags=""` `--optional-flags=''` and
`--optional-flags=[]`. We clearly need to move to another method for
building this command but for now we will just drop it.